+++ noodles-gff/src/io/writer/line/record/ty.rs
@@ -2,11 +2,14 @@ use std::io::{self, Write};
 
 use bstr::BStr;
 
+use super::percent_encode;
+
 pub(super) fn write_type<W>(writer: &mut W, ty: &BStr) -> io::Result<()>
 where
     W: Write,
 {
-    writer.write_all(ty)
+    let src = percent_encode(ty);
+    writer.write_all(src.as_bytes())
 }
 
 #[cfg(test)]
@@ -15,9 +18,18 @@ mod tests {
 
     #[test]
     fn test_write_type() -> io::Result<()> {
+        fn t(buf: &mut Vec<u8>, ty: &BStr, expected: &[u8]) -> io::Result<()> {
+            buf.clear();
+            write_type(buf, ty)?;
+            assert_eq!(buf, expected);
+            Ok(())
+        }
+
         let mut buf = Vec::new();
-        write_type(&mut buf, BStr::new("exon"))?;
-        assert_eq!(buf, b"exon");
+
+        t(&mut buf, BStr::new("exon"), b"exon")?;
+        t(&mut buf, BStr::new("exon%"), b"exon%25")?;
+
         Ok(())
     }
 }